Added Hide option within input window
Hide basically set the alpha of the root window to 0.5 when the focus has left kami. when the focus has returned it will set it back to 1. the option allows you to toggle this on or off within the input window and the default is set in the config.txt